+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/OnlineIndicator.kt
@@ -1,0 +1,27 @@
+package io.getstream.chat.android.compose.ui.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import io.getstream.chat.android.compose.ui.theme.ChatTheme
+
+/**
+ * Component that represents an online indicator to be used with [io.getstream.chat.android.compose.ui.components.avatar.UserAvatar].
+ *
+ * @param modifier Modifier for styling.
+ */
+@Composable
+public fun OnlineIndicator(modifier: Modifier = Modifier) {
+    Box(
+        modifier = modifier
+            .size(12.dp)
+            .background(ChatTheme.colors.appBackground, CircleShape)
+            .padding(2.dp)
+            .background(ChatTheme.colors.infoAccent, CircleShape)
+    )
+}
