+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/avatar/UserAvatar.kt
@@ -1,14 +1,10 @@
 package io.getstream.chat.android.compose.ui.components.avatar
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.tooling.preview.Preview
@@ -17,6 +13,8 @@ import coil.compose.rememberImagePainter
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.models.initials
 import io.getstream.chat.android.compose.previewdata.PreviewUserData
+import io.getstream.chat.android.compose.state.OnlineIndicatorAlignment
+import io.getstream.chat.android.compose.ui.components.OnlineIndicator
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
 
 /**
@@ -77,34 +75,6 @@ public fun UserAvatar(
         // Apply modifier to the avatar itself
         avatarContent(modifier = modifier)
     }
-}
-
-/**
- * Component that represents an online indicator to be used with [UserAvatar].
- *
- * @param modifier Modifier for styling.
- */
-@Composable
-public fun OnlineIndicator(modifier: Modifier = Modifier) {
-    Box(
-        modifier = modifier
-            .size(12.dp)
-            .background(ChatTheme.colors.appBackground, CircleShape)
-            .padding(2.dp)
-            .background(ChatTheme.colors.infoAccent, CircleShape)
-    )
-}
-
-/**
- * Represents the position of [OnlineIndicator] in [UserAvatar].
- *
- * @param alignment The standard Compose [Alignment] that corresponds to the [OnlineIndicator] alignment.
- */
-public enum class OnlineIndicatorAlignment(public val alignment: Alignment) {
-    TopEnd(Alignment.TopEnd),
-    BottomEnd(Alignment.BottomEnd),
-    TopStart(Alignment.TopStart),
-    BottomStart(Alignment.BottomStart)
 }
 
 @Preview
